[bphh-216] Refactor number requirement unit selector so that a no unit option has a value of undefined instead of set enum option
