+++ app/frontend/components/domains/requirements-library/requirement-field-edit.tsx
@@ -23,8 +23,7 @@ import { Controller, FieldValues } from "react-hook-form"
 import { FieldPath } from "react-hook-form/dist/types"
 import { UseControllerProps } from "react-hook-form/dist/types/controller"
 import { useTranslation } from "react-i18next"
-import { ERequirementType } from "../../../types/enums"
-import { ENumberUnit } from "../../../types/types"
+import { ENumberUnit, ERequirementType } from "../../../types/enums"
 import { EditableInputWithControls, IEditableInputWithControlsProps } from "../../shared/editable-input-with-controls"
 import { UnitSelect } from "../../shared/select/selectors/unit-select"
